Daily LeetCode (11/22)
Decent Daily, I struggled with the problem because my initial solution
checked alot of positions in the matrix that don't exist. The final
solution that I'm commiting uses the fact that any elements indexs in
the matrix (i + j) add up to the nth diagonal that they are apart of. We
used a hashmap with i + j as the key and then simply reverse each array
and add it to the answer. (overall pretty easy brute force solution, but
hard to find a good solution imo.)