Daily LeetCode (11/28)
Really easy Daily. I basically got the combinitorics solution without
hints. I did get stumped for by not realizing you just have to check if
the chair count is still 2 at the end of the function. Other than that,
it seemed rather easy, and I imagine there is a much harder DFS solution
that made it a HARD problem.