Daily LeetCode (11/30)
Extremly difficult problem for me. Generally anything with bits
imediately scares me. I didn't solve this by myself and eventually had
to look up the solution, but it mostly makes sense now. Still not 100%
sure I could come up with this solution if I was given this question in
an interview a year from now.