Advent of Code (Day #10)
Today's challenge was actually pretty difficult. My solution for part
one was much more verbose then it needed to be, though decently
efficient. For part 2 I came up with a witty solution that essentially
checks how many walls are to the left of all non main path nodes. There
were some annoying edge cases that made the problem take a bit longer,
but overall this was the best placements I've had so far!